Change page-footer secondary_button params
The current params set 2 things but, based on
current usage, only used to make the footer button
into a `<a>` element with secondary button
styling.

Having these 2 params also leads to complicated
logic in the template.

This reduces them to one and uses it to infer the
type of element and styling.

If we need things possible but unused with the
current params in the future, like making a
link-button with primary styling, we can always
revert or extend these changes but for now this is
the most compact option.