make sure `user` is set when calling the permissions template
the template can be called from invite-user.html or
edit-user-permission.html. make sure if we call from invite user (where
`user` is not set), there is a user variable with flags set such that
all the correct parts of the page do/don't display. obviously a newly
invited user wont be platform admin or have webauthn yet, but they need
a truthy mobile number or it'll disable the sms auth option thinking
it's editing an existing user with no mobile number