Re-crop illustrations to fit in 1/2 column
Since we have 2 choices on the page it is possible to have the grid be
two columns, rather than the three we had in the previous incarnation of
this page.

This gives the text labels the most space to display without wrapping.

In order for the 2 column layout to look good we need to:
- make sure the images don’t get zoomed in too much
- make sure the images don’t take up too much vertical space and push
  the radios down off screen on smaller laptop displays

To do this I have:
- changed the aspect ratio of the image to the design system default
  3/2
- preserved the size of the logo/crest as it would have been in a 3
  column layout and only widened the grey bard which represent the text
  of the email