Change org settings page to show "Manage" not "Change" branding
We want to start using the term "Manage" when you have the option to
change or manage more than one thing, for example email reply-to
addresses. This changes the branding rows in the org settings table to
fit the new pattern. Other parts of the settings table and app will be
looked at later.