simplify the govuk_checkbox_field_widget function
there was some logic based on how it handled wrap_in_collapsible -
conveniently that's only set by the GovukCollapsibleCheckboxesField
class, so we can just handle that logic in that class.

Also, make use of the `param_extensions` we check for everywhere by
setting the collapsible specific params in that variable.

the govuk_checkboxes_field_widget function was then only being called
from GovukCheckboxesField.widget, which makes sense - so we can just
remove the function and move its contents in there.