make checkboxes extend GovukFrontendWidgetMixin
now we've done the hard work to massage the widget function into the
right shape, it's just a matter of moving the
prepare_params_for_govuk_checkboxes function on to the class and making
its function signature match.

Note that the CollapsibleCheckboxesField still overrides widget, but it
calls super().widget, and then wraps the result, which I think feels
quite tidy and reflects the changes it makes to the normal checkboxes
class.