move all text inputs to use GovukFrontendWidgetMixin
move the contents of govuk_text_input_field_widget to
`GovuktextInputFieldMixin.prepare_params`, and make the `type` field
into an input_type member variable of some of the classes. It defaults
to `text`.

Note that the prepare_params function now refers to self rather than
field, and sets `params.type` to equal `"text"` by default. Previously
it didn't set `params.type` at all, and relied on govuk-frontend-jinja
to set the type to be text if unset, but I thought being explicit helps.