Generate param_extensions in init
We are editing a dictionary in the __init__ method for
EmailBrandingChooseLogoForm, which is stored on the class. This can lead
to data being mutated between requests on the same worker.

Let's build a whole new param_extensions dict every time we instantiate
the form, rather than injecting into an existing dict.