fix fields sharing param_extensions between them
this is because param_extensions is defined once at a class level, and
then we're copying things to it. Part of the confusion stems from the
Field not actually be instantiated when you call `my_field =
Field(...)` (that just creates an UnboundField object), but when the
form is created. By this point, any other forms that have been created
will have a copy of the exact same Field.param_extensions object, and
modifying one dict will modify all of them.

You can have a play about by commenting out the deepcopy and playing
around with the unit test provided.