Prepare to bump Flask dependencies
Flask 2.2.0 deprecates the _app_ctx_stack and _request_ctx_stack. We
need to migrate away from using these in order to prepare to upgrade our
libs to the latest versions.

The recommended place to store app/request data is now on the `g`
object, see https://flask.palletsprojects.com/en/2.2.x/api/#flask.g for
info.

When we are using the ServiceAPIClient during tests, we aren't inside a
request, so `g.current_service` hasn't been populated. We don't need to
run the `check_active_service` logic and aren't testing for that here,
so let's just patch it out.