Use org model to update an org's branding, and clear cache if necessary
Intead of using the organisation client to update an organisation, this
uses the method on the organisation model. When we change the branding
of an organisation from no branding (or GOV.UK) to a set brand, this
affects all of the organisation's services so we need to clear the cache
of those services too. This change passes in the argument to the
`update` method to make sure that we clear the cache for services when
necessary.