Add extra logo validation and manipulation
Let's run some extra checks on the uploaded logo to try to make sure
it's of good enough quality to accept.

Specifically, we:
1) Make sure it's at least 108px tall as an indicator that it is
high-quality enough.
2) If it's wider than our max width, we scale it down (respecting the
aspect ratio)
3) If it's not <108px tall, let's pad the top and bottom with
transparency.

We add a new `image_processing` module to abstract the use of PIL and
limit the image manipulation interface to only the things we actually
need/expect to use.