DRY GovukRadiosWithImagesField param_extensions
Move the bits we can to class level to avoid duplication.

The bits that might change based on the service, call from the template
when rendering. This means we no longer need to pass in a service to the
constructor, which is nice. This is consistent with how we generally
call form constructors - the only things that go in are things that
directly influence field choices (eg, a list of options from the
database to show as checkboxes)