pass webauthn_credential_id back to the api when logging a user in
as a refresher, we pass all the user's credentials to webauthn_server,
where Fido2Server tries each credential sequentially to see if it
matches the webauthn key that just tried to log in. If it finds one
succesfully, then it returns that cred's AttestedCredential. We have to
then loop through the credentials again to compare all the user's creds
to the one that matched, to get the ID for that credential.

Then we can pass that back in to the webauthn complete api call. If that
api call is succesful (ie: the user hasn't been locked out for too many
tries) then we'll set the webauthn_credential.logged_in_at via the api.