Refactor generation of grid column classes
Redefining $govuk-grid-widths again is a bit
brittle because we're constantly having to check
our version for parity when we update. This moves
to us adding our custom values to the existing map
instead.

This means the worst thing that can happen is that
one of our values overwrites the same pair which
should only result in a subtle change to the
percentage used.

I've checked and this still outputs the same
classes we have now.

Article on Sass maps, for reference:

https://sass-lang.com/documentation/values/maps#add-to-a-map