add alt_text to email branding
This is always required at the email_branding_client level, however,
rather than add form fields for it, for now we'll just copy the name
across. We don't want to confuse ourselves - it's not currently being
used anywhere, and currently we just use the branding name as the alt
text.

After we've migrated existing data across, we can then start using it in
the email html template, at which point we should expose it in the edit
form and let it be set separately to the name.