Add page to preview a new email brand
Implements the basic view for previewing a new email brand. This page
uses the existing `email_template` endpoint that can take in a set of
parameters used to define the branding banner that should appear at the
top of the email.

The new view doesn't handle pressing the button to create the brand, as
we are now moving the "name logo" input to this page. It's not possible
to create the brand without the name, so this page simply shows the user
what the brand looks like.