Fix back_links when dropping out of new brand flow
If a user bails out of the new email branding flow, we take them to the
page where they can submit a zendesk ticket with details on the brand
they want. This page has a back link, and it should take the user back
to the correct place in the previous flow. We need to pass through URL
query params and the correct back links to build the right URL.