Import GOVUK Frontend JS as real ES Modules
GOVUK Frontend, from v4.1.0, allows us to import
each component's JS as a real ECMAScript (ES)
Module.

Note: we ship our JS as a single JS file but part
of this is made of JS, written as ES Modules and
bundled together using RollupJS.

We were previously importing the component JS in a
slightly hacky way. Until v4.1.0, GOVUK Frontend
made its JS available as a single file, containing
the JS for all components. It did, however, also
ship the source code of all this JS in the NPM
package. Each ones of these files is written as an
ES Module, to make dependencies clear and to make
it easy to bundle them. We only wanted to include
the JS for the components we use so we were
importing these files, despite this code not being
officially supported by the Design System team.

v4.1.0 makes all the GOVUK Frontend code available
through an API that lets you import either the
whole lot or just the code for the components you
use. These changes rewrite our JS to use this API.

Because the GOVUK Frontend package now exports all
the component JS as part of its module API, and
the version of NodeJS we use to build our JS
supports ES Modules, we no longer need to use the
commonJS plugin to load GOVUK Frontend into
RollupJS (our bundler). This commit therefore
removes it.