Make other skiplinks work with GOVUK Frontend JS
We have a few in-page skip-links which can be
enhanced with the same JS used for the one at the
top of the page.

This also includes a fix, correcting a skiplink
which linked to an element that didn't exist
(id="content"). This assumes the intention was for
it to link to the main content area.