Make `brand_type` of GOV.UK brand `"govuk"`
The GOV.UK brand is a special case because it doesn’t have an `id`.

It’s inconsisent that when creating a model object for it the
`brand_type` column is set to `None`. This commit fixes that by setting
the `brand_type` to `"govuk"` in this special case, and using that as
the canonical field for whether or not a brand is GOV.UK.