Hard-code the HTML for the 2-factor webauth error
We used the GOVUK Frontend error message component
for this. Recent updates to our version of GOVUK
Frontend changed the elements used for this
message from `<span>`s to `<p>`s. This broke our
code, meaning part of it is now showing.

This hard-codes the original HTML, to unblock
deployment of a recent bump to the latest GOVUK
Frontend until we can fix it properly.