add alt_text field to platform admin manage branding page
add validation to only allow one of alt_text and text.

follow the existing pattern of checking the operation that is passed in
from which submit button the user clicked to make sure this validation
doesn't trip if the user is uploading the logo.

also fix the tests.
moving to populating the form from obj required us to mock out some more
things that we didnt need to mock out before :thinking_face: