Give pill item link text new underline styles
You can't override `text-decoration` styles set by
a parent on children, but our pill item links set
different values for their number and its label.

See: https://developer.mozilla.org/en-US/docs/Web/CSS/text-decoration

Because of this, we need to turn `text-decoration`
off on the parent and assign it individually on
the label (which is underlined by default and
should get the hover style).

This adds the `govuk-link--inverse` class to links
in pills to make them white in all states.

The new link styles remove the need for a change
in colour when hovered as the underline becoming
thicker replaces this as a visual hint.

A nice byproduct of these changes is the removal
of the need for the `govuk-link--no-visited-state`
class. The :visited state is one of those
`govuk-link--inverse` sets as white so we don't
need to set that ourselves.