Give dashboard count link text new underline style
You can't override `text-decoration` styles set by
a parent on children, but our dashboard count
links set different values for their number and
its label.

See: https://developer.mozilla.org/en-US/docs/Web/CSS/text-decoration

Because of this, we need to turn `text-decoration`
off on the parent and assign it individually on
the label (which is underlined by default and
should get the hover style).