add error messages and change names of the field
lead with the assumption that we'll be thinking in terms of "do i want
text on this branding", and alt text comes from that decision, rather
than being the thing you lead with.

also call it logo text rather than just text to be a bit less ambiguous.
similar to what we say in the gov identity flow