Rewrite show-more component to use new link style
This currently fakes the link style with a
border-bottom, which won't allow us to mimic the
new link style.

These changes rewrite the component so the link is
real and so can benefit from the new styles.

Also includes some CSS to ensure the link has the
same focus state (a 3px outline) in forced-colour
modes.

Apart from the changes to the show-more CSS, this
commit also removes an unused import for the
show_more macro.