Use Design System text input prefix styles
Update the hex colour code picker component to support GOV.UK Design
System text input prefixes

https://design-system.service.gov.uk/components/text-input/#prefixes-and-suffixes

This will mean that users never need to enter the `#` for their hex code
values. But we still want to make sure that value is present in the
database, so we do some magic to:

1) Add it to any submitted value via `process_data`
2) Strip it whenever we render the value in HTML via the `_value`
method, which we make sure is used in our Govuk

We need to adjust the colour picker JavaScript to append the colour
swatch to the right element, as the Design System preview adds a wrapper
div around the prefix and input if applied.