Make colour picker work when there is whitespace
Sometimes when you copy and paste a hex code from another place you end
up with leading and/or trailing whitespace.

Our Python code will handle this automatically and strip the whitespace
before sending the data to the API:
https://github.com/alphagov/notifications-admin/blob/main/app/main/forms.py#L476-L490

This commit makes the Javascript strip the whitespace before using the
value of the hex colour preview.