Convert edit-letter-template links to new style
Add `govuk-link--inverse` class to
edit-letter-template links to make them white in
all states.

The new link styles remove the need for a change
in colour when hovered as the underline becoming
thicker replaces this as a visual hint.

A nice byproduct of these changes is the removal
of the need for the `govuk-link--no-visited-state`
class. The :visited state is one of those
`govuk-link--inverse` sets as white so we don't
need to set that ourselves.