Change classes for show-more variant
The changes in this pull request assumed the
variant of the show-more component where it has no
line running through it had these two classes:
1. `show-more`
2. `show-more--no-border`

It instead had just one: `show-more--no-border`,
which meant the CSS was applied incorrectly and
made the link fill the viewport(!).

This rewrites the macro so both classes are
applied so the CSS works as expected.