Allow inbound sms to be turned on by users
The "Receive text messages" option on the service settings page used to
show a page with information and a link to contact support. We want to
make this self-service. Choosing that option now shows you a page with
information, followed by a page to turn the feature on. It's not
possible to make turning off inbound sms self-service so if someone
clicks the option to turn off the feature we show them a page which
explains that they don't need to use the inbound number and also gives
them the option to contact support if they want it removed from the
service permanently.

The routes have been renamed, but this temporarily keeps the
"/set-inbound-sms" endpoint so that people don't see an error if they
are in the middle of trying to use this feature as the change is being
deployed. It will be removed later.