Link to templates from main nav for org users
We want to let organisation users approve go live requests for teams in
their organisation.

As part of reviewing these requests, we think that being able to see the
templates a team has set up will be important.

This commit makes ‘Templates’ visible in the left hand navigation when
an organisation is viewing a service.