Give org users permission to view templates
We want to let organisation users approve go live requests for teams in
their organisation.

As part of reviewing these requests, we think that being able to see the
templates a team has set up will be important.

This commit makes the various pages under the ‘Templates’ menu visible
to someone in an organisation viewing a service.