Clear branding pool caches when updating branding
If we update email or letter branding, we need to clear the branding
pool caches to avoid them still containing the old branding details.
Finding out which branding pools (if any) contain the branding we're
updating would require an extra api request, so this just clears all the
branding pool caches.