Be stricter at looking for gov identity example
When we show an example of government identity branding we try to find
the DfE one from the database.

Because there are several brands on production with "Department for
Education" in the name this doesn’t always match the right one.

This commit makes the matching stricter, so that the right one will be
shown.