Cache email_safe
We’re calling this is a loop with the same value repeatedly. Not sure if
it’s slow enough to be noticeable, but feel like it can’t hurt to cache
it since string manipulation can be kind of slow.

Chose 4 because you’d call it twice to do a comparison, then double that
for some headroom.