Validate HTML as part of running the tests
HTMl is a forgiving language, so valid markup is not always needed in
order for a page to render properly.

But invalid markup can:
- cause problems with spacing or layout where elements are improperly
  nested
- cause problems for users of assistive technology, if element names are
  mispelled (for example `<thread>` will be ignored whereas `<thead>` has
  semantic meaning)
- by symptomatic of code that is hard to follow because it’s not
  indented properly

This commit adds a step to `client_request.get()` which checks for valid
HTML. If invalid HTML is found then an exception will be raised.

It only runs on HTML documents. Almost all of our pages are HTML, the
only exception being the email template, which is nominally XHTML (and
hard to validate because of all the tricks we use to get around email
client rendering issues).