Show divider if > n options for branding
If there are lots of items in the pool, it’s handy to flag that
_Something else_ is a separate, special option, without making you scan
the whole list.

We can do this by putting a divider ahead of it.

We don’t need to do this when the list is very short, which is also the
case where there are no items in the pool and all the choices are magic
options.