Add functionality to put divider between radios
The radios component from the GOV.UK Design System supports putting
text dividers between radios. This is most commonly a way to show that
the last option is some kind of alternate or fallback, but they can be
placed at an arbitrary point in a series of options.

The JSON required to render one of these is `{"divider": <text to be displayed>}`
https://design-system.service.gov.uk/components/radios/#radio-items-with-a-text-divider

This commit adds a class (`GovukRadiosField.Divider`) which can be used
instead of a regular `tuple` of `(value, label)`.