Change webauth pages use GOVUK error summaries
They currently use a weird mashup of our banner
component HTML and the GOVUK error message
component styles.

This changes them to instead use the GOVUK error
summary component.

Note: I've also moved the error summaries to the
top of the page to follow design system guidance.

Includes a potentially contentious change to the
test that checks paragraphs for the `govuk-body`
class to let error summaries not trigger it,
because it contains paragraphs without that class.