Bump utils to 60.0.0
This brings in a breaking change to how pyproj transforms polygons. I've
removed pyproj as a top level dependency in the requirements.in file to
avoid version conflicts - we only had it as a top level dependency to
check it could be installed on the PaaS when we were using an older
version of Python.