escape user-entered values in banner
specifically this can be seen when deleting a template - if you name
your template `<script>alert(1)</script>` for example, that html will be
rendered and the alert will happen

jinja treats all variables as unsafe unless we declare them safe (Note
that all jinja macros are considered to return fully formed HTML, thus
are declared as safe internaly automatically). To mark a variable as
containing HTML you wish to render, you pass it through the `safe`
filter `{{ foo | safe }}`. This is what the govuk-frontend-jinja does to
any html you pass in through the html property.

When showing a delete prompt we use this html field to wrap the name of
the thing with a visually hidden span so it is read out by screen
readers but not visible to the user (we do this because the name's
already above in the `body` variable normally).

However, we had an XSS vulnerability where the name of the template
isn't escaped, and then it's marked as safe. It's not safe, as it's
untrusted user input.

To fix this, rather than concatenating the string in python inside the
macro invocation, we can create the string in html in a set block above.
By calling `{{ thing }}`, jinja automatically escapes any HTML contained
within.