avoid concatenating html in python
nothing i've found here has had untrusted content, but that doesn't mean
we couldn't call these with user content in the future, so move them all
to using jinja `{{ foo }}` syntax to ensure content is always escaped
(see prev commit for more details).

as a rule of thumb, all times we pass `html` in should be called using a
`{% set html %}...{% endset %}` if we're adding variables in.