protect against XSS in the identity branding logo text hint text
same principles as before about moving format strings to the template.
However, this fixes an additional bug! Because we were overwriting the
format string with a formatted string, this would mean that the second
time that specific python process loaded a form, the string would have
the previous organisation's name rather than the current org.

moving the string to the template both protects against XSS and also
fixes this issue