Remove unused code path
The code in this view is potentially called in three places:

1) Directly
2) Via `get_notifications_as_json`
3) Via `view_notifications`.

In cases 2 and 3, the CSV logic won't execute because the request
path doesn't end with '.csv'.

In case 1, there are no other places in the admin codebase that generate
URLs to this view which indicate it's currently unused. Checking Athena
for requests in the last year also shows no hits to this endpoint.