create route for uploading a letter branding
platform admins get there if they chose "something else". Unfortunately,
on this page there's a text box for users to describe what they want,
which gets fed in to a zendesk ticket.

For now, rather than re-engineer the form to choose whether to show that
field based on if the user is a platform admin or not, just keep the
field and require that platform admins have to put in a value.

note, we're not calling this a logo, since it is the entire branding
(while in email flow there is the logo which, along with layout and alt
text makes up the branding)