Add test tracking all app routes
Whenever we remove a route from the admin app, we should think about how
this might impact on any users who've saved that route (eg in a
bookmark) and try to put a redirect in place.

This patch adds a test to remind us to think about this. It checks
against a cached list of all the admin app routes. We add a helper
command to generate this automatically so that the dev burden is low.