Handle nested `<tbody>` tags more elegantly
We have to avoid nesting `<tbody>` tags because it’s not valid HTML.

The hacky way around this was to close the outer tags before opening the
inner ones, leaving an empty element.

This commit changes that by adding a flag to control whether the outer
ones are printed at all.