Style navigation link differently
So that it stands out from the rest of the links in the navigation, and
so that it feels like a different/separate thing.

Implementing this meant removing the targetting of the `li` element
in the CSS, because this overrides any class-based targetting. So that
we can use class-based targetting to style our new navigation element,
we need to use class-based targetting for all navigation elements. This
is in line with the naming conventions in GOV.UK Frontend anyway.