allow previewing of arbitrary svgs in letter preview
previously it took a letter_branding uuid and got the filename from the
database.

now, allow user to pass in filename straight away if it's not
necessarily been saved to the database yet.

don't allow the user to pass in both.

Also tweak the existing branding preview macros to make them a bit more
DRY.