Add NewRelic to dev+preview
We add basic APM instrumentation for NewRelic. This will inject tracing
headers that filter through to any other instrumented apps and send them
to NewRelic so that we can see how requests travel through our services
and how long each span takes. This should help us highlight slower areas
of the codebase which we can then investigate and optimise.

We only send APM information - not logs - here, and we heavily limit the
information sent so that we hopefully aren't gathering much/any PII. We
also only deploy to dev and preview at the moment, so there's no risk of
this code gathering real user data for now. This will allow us to safely
investigate the tool and see whether it's something useful for us.