Only show ‘not set up’ message when user has chosen organisation branding
If the user has chosen their organisation’s branding but we don’t have
it then we should let them know that they’re going to need to set it up.

I can’t remember if we did this and it changed to talk about the
service[1] or if it was never like this.

Anyway, talking about the service’s branding not being set up doesn’t
make sense when the user has pick an option labelled with the name of
their organisation, so let’s fix it.

1. `git blame` got me back as far as https://github.com/alphagov/notifications-admin/commit/7dafb91fb3bca38248236849e0b6b9fce7ecada4#diff-626679b981f01eb2e5471518344e7d131ef28b01b8d0aafbf2bc31d2f704ccabR35-R39 which doesn’t explain why the line was added