Remove template lists from `<nav>`s
We removed the `<nav>` around template list items on the templates page,
but there are a couple of other places where we show a template list
that still needed updating - the copy template page and the page to
choose a template to reply to a received text message with.

Changing the tag used doesn't affect the appearance of the pages.