Let organisation users see all template folders
We already let organisation users see all templates in their service:
https://github.com/alphagov/notifications-admin/pull/4494/files

However if the template is in a folder then the user won’t be able to
see it, since the user also needs permission to see the folder in order
to see the template.

This defeats the objective of letting organisation users see all
templates.

This commit effectively gives organisation users permission to see all
folders in a service, and by extension all templates no matter which
folder they are in.