Remove unused Jinja macros
This uses the script from https://github.com/alphagov/notifications-admin/pull/4534/files
to identify Jinja macros which aren’t imported anywhere.

It then removes them.

There were a couple of false positives, namely:
- `govukRadiosWithImages`
- `govukCheckboxes`

It’s hard to programitically identify where these are because of the way
we’ve wrapped the GOV.UK Frontend components, but it is happening:
https://github.com/alphagov/notifications-admin/blob/8236a28e9aff6388857335aca388e04a3bcdfdf4/app/utils/govuk_frontend_field.py#L123