Show whichever daily limit is the constraint
Right now services have an overall daily limit of messages.

We are instead going to give them a per-channel limit. This per channel
limit is currently enforced, but set to an impossibly high number
(999,999,999).

As part of the migration we will:
- increase the overall limit to an arbitrarily high number
- decrease the per-channel limits to a lower number, which will become
  the contraint on how many messages a service will send

This commit implements logic such that the page works both before and
after the migration.