remove check_inactive_service from api calls
we added this as a safeguard to prevent people modifying inactive
services. However, when a user hits an endpoint that is associated with
a service we decorate that service with the `user_has_permissions`
decorator.

Among the various checks the `user_has_permissions` decorator does is
`user.belongs_to_service` - this checks in the list of services returned
from the api in the get_user_by_id blob. In the API we explicitly only
return services that are active in that list[^1]. It also later checks
that any permissions match (if that endpoint requires specific
permissions to hit for that service). Again, the API explicitly filters
to only return permissions for active services when returning the
user[^2].

Note that platform admins bypass both these checks (unless
`restrict_admin_usage` is set).

This means that there's no way a user can ever enter the view code at
all with an inactive service, be it for a GET or a POST. The only risks
are if we don't clear the user cache when updating a service - however
we explicitly clear user cache when archiving a service[^3].

This means we can remove the check_inactive_service, as we're confident
this code is never tripped legitimately. The only time it can ever be
reached is if the user is in an endpoint that has a different decorator
(specifically `@user_is_logged_in`) so doesn't look at service
permissions etc. Previously we were blocking POST updates on pages with
these decorators, and causing bugs where users couldn't create new
services or update their name/password/etc if they'd recently viewed an
archived service, but now we have fixed that bug properly.

[^1]: https://github.com/alphagov/notifications-api/blob/ef555db14afa96ba0122341154b00d14edf9a98d/app/models.py#L179
[^2]: https://github.com/alphagov/notifications-api/blob/ef555db14afa96ba0122341154b00d14edf9a98d/app/dao/permissions_dao.py#L65-L71
[^3]: https://github.com/alphagov/notifications-admin/blob/b6a16bebea65ca59360261b1b1ee8aab58e54897/app/main/views/service_settings/index.py#L382-L386