Check branding not in use before archiving it
If it is set as a service branding for an active service or if
it is set as a default branding for an active organisation,
we do not want to archive it.

If it's a part of a branding pool, but not currently used by any
service or organisation, we are happy to archive it (and
we remove it from any pools as a part of archiving process in the
API).