Try uploading new branding file before saving to database
We were trying to save the new letter branding details to the database
first then attempting to upload the file to S3. If uploading failed, we
then rolled back the changes to the database by calling the
`update_letter_branding` client method again with the original branding
details. Now that we also store an `updated_by_id` and `updated_at` value
for letter branding this is harder to do - we don't have those values in
admin.

This changes the order of the code so that we attempt to upload the file
to S3 before saving to the database. If saving fails, there are no
database changes to rollback. It is possible that the database update
fails because the logo name is not unique, which will leave us with an orphaned
logo file in S3. However, this would be fairly rare and shouldn't case
any issues. We can look at doing making the the validate method on the
form check if the name is unique later.