Use `current_user` in the branding files
Some of the old code for email branding and letter branding was getting
the id of the current user from the session using `session["user_id"]`.
We can use the `current_user` proxy instead, which is more consistent
with our code in other places and saves some set up in the tests.