Rename branding files that have been created by users
This renames the logos to remove the temporary prefix from the start to
match what we do when a platform admin user uploads a logo. If we don't
rename the logos there is a chance that any logos that a platform admin
user uploads using the user-facing routes will get deleted. This is
because we delete all logos with the temporary prefix that a user has
uploaded when they use the platform admin branding routes.

This is a temporary fix to stop us creating any more logos which have a
"temporary name" (and are at risk of deletion). The longer term plan is
to use S3 tags to indicate which uploaded logos aren't being used for
branding and can be deleted.