An admin page to view a single email branding
When you go to a list of all brandings, when you click a branding link,
that used to take you straight to "update branding" page.

Now that we can archive brandings, and if they belong to orgs or services
they can't be archived, it's useful to also have a page where admin
can view the branding and what orgs / services it is associated with.

Also seeing a branding preview is useful if you are about to edit it.

This page has:
- branding preview
- a list of orgs that have the branding as a default option
- a list of services that use the branding
- edit link
- archive link