Update features page
Update our features pages to complement the new 'Using Notify' guidance
we've publishied.

Explanatory content has moved to the 'Using Notify' pages, so we're
paring the Features pages back to a single page with a list of
highlights. Users can click on the links for subjects of interest to
discover more.

Changes:

* update the content for the main features page, including links to
  Using Notify and Pricing pages
* delete the features sub-pages for email, text messages and letters
* update routes and navigation subdictionaries

Some link text or destination page headings may need to change, but that
will come in a later PR