Stop updating `message_limit`
When we make a service live it should already have a very high
`message_limit` so we don’t need to increase it.

When we put a service back into trial mode we shouldn’t decrease its
message limit, because then this decrease would stick when making it
live again.