Make one true `IntegerField`
Rather than having two different types of `IntegerField` let’s settle on
one.

Going with the existing `GovUkIntegerField` name because this is what
we reached for last time we wanted to use an `IntegerField`.

This required a bit of finangling to make the previously-named
`ForgivingIntegerField` work when it’s not bound to a form.