Update platform_admin_create_email_branding
Switch this view over to using the new S3 logo_client interface. One
notable change we make here is that we also upload the new logo to its
permanent home in S3 before creating the logo in the DB. I think it's
better to have the logo there in advance, than potentially create an
email brand in the DB and then fail to upload the logo.