Update FileStorage stream
When files are uploaded, WTForms creates a
werkzeug.datastructures.FileStorage wrapper around the stream. This
provides some extra metadata about the file that was uploaded, such as
the content type and filename, which we use to pass through to the logo
uploading calls.

The user-uploading logo validation functions manipulate the stream to
make sure it's the correct size, so we need to replace the user's stream
with our updated one.