remove second hr from view branding page
when a branding doesn't have created_by or updated_at set, it would show
two horizontal rules, which looks strange. And there isn't as much a
need to draw a clear line between the updated at text and the
edit/delete buttons, as they look very dissimilar. It's useful to still
have an hr if the last line of the main body is a link to a service that
uses it, as we want to clearly separate that link from the edit/delete
links