Use tagging rather than path-based temp logos
Our letter logos have paths hard-coded in our utils templates to render
correctly, and fixing this is going to take more dev effort than we want
to expend right now.

Let's switch to using tagging-based lifecycles for temporary logos
instead, which means we don't need to tackle and migrate letter logos
and can store temporary logos anywhere. It will be less immediately
obvious whether a logo is temporary or not, but allows us to move
forwards and still deal with the initial bugs that kicked off this work.